CSharp/WorkerRole1/requirements.txt
azure-servicebus==0.20.3
azure-storage==0.33.0
bin\Pillow-3.3.1-cp34-none-win32.whl
bin\cffi-1.8.2-cp34-none-win32.whl
bin\cryptography-1.5-cp34-none-win32.whl
Flask==0.11.1
flask-mongoengine==0.8
Flask-WTF==0.12
idna==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.6
pyasn1==0.1.9
pycparser==2.14
pymongo==3.3.0
python-dateutil==2.5.3
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
WTForms==2.1